Added InProgress build status, worked around subtle date/time issue in PollingService causing missed build notifications
Added an additional status of InProgress - if the build.Result is null and InProgress = true
Changed the PollingService logic so that it is only looking for either new builds or builds with a changed InProgress indicator. This removes the reliance on synchronised date/times between the local machine and the TFS build machine which was causing completed build notifications to be missed.